+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetitemop.py
@@ -1290,7 +1290,7 @@ class Evaluations3(BaseModel):
 
     source: Optional[CreateDatasetItemEvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2025-10-10T04:39:50.168Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2025-10-10T11:07:58.032Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1342,7 +1342,7 @@ class Evaluations2(BaseModel):
 
     source: Optional[EvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2025-10-10T04:39:50.167Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2025-10-10T11:07:58.032Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1394,7 +1394,7 @@ class Evaluations1(BaseModel):
 
     source: Optional[Source] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2025-10-10T04:39:50.167Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2025-10-10T11:07:58.032Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1468,5 +1468,5 @@ class ResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-10T04:39:40.255Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-10T11:07:48.042Z")
     r"""The date and time the resource was last updated"""