+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetop.py
@@ -83,5 +83,5 @@ class CreateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-10T04:39:40.255Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-10T11:07:48.042Z")
     r"""The date and time the resource was last updated"""