+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -230,7 +230,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K766P1NS4TVJMAJ0G311KP7R"
+        "01K76WWQR1D6XS7KA5TTXKSS1S"
     )
     r"""The id of the resource"""