+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K766P1NSFFA9QE1XTBWPWBE9"
+        "01K76WWQR1SC2FYVXA65ZXZ1K9"
     )
     r"""The id of the resource"""