+++ packages/orq-rc/src/orq_ai_sdk/models/updatebudgetop.py
@@ -208,5 +208,5 @@ class UpdateBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-10T04:39:40.255Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-10T11:07:48.042Z")
     r"""The date and time the resource was last updated"""